pyproject.toml
[tool.poetry]
name = "moneyhand"
version = "0.1.0"
description = "Moneyhand - To keep your finances in order."
authors = ["Nikita Kharlov <nikita@harlov.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
asyncpg = "^0.21.0"
click = "^7.1.2"
alembic = "^1.4.3"
psycopg2 = "^2.8.6"
nest-asyncio = "^1.4.1"
pydantic = "^1.6.1"
appdirs = "^1.4.4"
aiogram = "^2.10.1"
matplotlib = "^3.3.2"
fastapi = "^0.61.2"
uvicorn = "^0.12.3"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = {version = "1.4.0b1", allow-prereleases = true}

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
pytest-asyncio = "^0.14.0"
black = "^20.8b1"
docker = "^4.3.1"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
